[warning-fix][signed-unsigned-mismatch] Fix warning oce/src/Image/Image_PixMap.cxx:410:0 Comparison between signed and unsigned integer expressions

+++ src/Image/Image_PixMap.cxx
@@ -407,8 +407,8 @@ Quantity_Color Image_PixMap::PixelColor (const Standard_Integer theX,
                                          const Standard_Integer theY) const
 {
   Standard_Integer aScanlineId = myImage->getHeight() - theY - 1;
-  if (theX < 0 || (unsigned int)theX >= myImage->getWidth() ||
-      theY < 0 || (unsigned int)theY >= myImage->getHeight())
+  if (theX < 0 || theX >= myImage->getWidth() ||
+      theY < 0 || theY >= myImage->getHeight())
   {
     return Quantity_Color (0.0, 0.0, 0.0, Quantity_TOC_RGB);
   }
